created new Gamemanager script
