+++ Assets/GameManager.cs.meta
@@ -1,8 +1,11 @@
 fileFormatVersion: 2
-guid: 5f1c8695fb4994c8d9f06b603b97fda6
+guid: 58a6e0c3b12264c3b84164b278ae50e4
 MonoImporter:
+  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
   icon: {instanceID: 0}
   userData: 
+  assetBundleName: 
+  assetBundleVariant:
